requirements.txt
mysql-connector-python==8.0.20
numpy==1.18.4
protobuf==3.12.0
six==1.14.0
